+++ lib/pg_eventstore/commands/multiple.rb
@@ -4,8 +4,8 @@ module PgEventstore
   module Commands
     # @!visibility private
     class Multiple < AbstractCommand
-      def call(&blk)
-        queries.transactions.transaction(&blk)
+      def call(read_only:, &blk)
+        queries.transactions.transaction(read_only: read_only, &blk)
       end
     end
   end