+++ lib/pg_eventstore/queries/transaction_queries.rb
@@ -23,24 +23,30 @@ module PgEventstore
     end
 
     # @param level [Symbol] transaction isolation level
+    # @param read_only [Boolean] whether transaction is read-only
     # @return [void]
-    def transaction(level = :serializable, &blk)
+    def transaction(level = :serializable, read_only: false, &blk)
       connection.with do |conn|
         # We are inside a transaction already - no need to start another one
         next yield if [PG::PQTRANS_ACTIVE, PG::PQTRANS_INTRANS].include?(conn.transaction_status)
 
-        pg_transaction(ISOLATION_LEVELS[level], conn, &blk)
+        pg_transaction(ISOLATION_LEVELS[level], read_only, conn, &blk)
       end
     end
 
     private
 
     # @param level [String] PostgreSQL transaction isolation level
+    # @param read_only [Boolean]
     # @param pg_connection [PG::Connection]
     # @return [void]
-    def pg_transaction(level, pg_connection, &_blk)
+    def pg_transaction(level, read_only, pg_connection, &_blk)
       pg_connection.transaction do
-        pg_connection.exec("SET TRANSACTION ISOLATION LEVEL #{level}")
+        if read_only
+          pg_connection.exec("SET TRANSACTION ISOLATION LEVEL #{level} READ ONLY")
+        else
+          pg_connection.exec("SET TRANSACTION ISOLATION LEVEL #{level}")
+        end
         yield
       end
     rescue PG::TRSerializationFailure, PG::TRDeadlockDetected