+++ spec/pg_eventstore/subscriptions/command_handlers/subscription_runners_commands_spec.rb
@@ -9,20 +9,14 @@ RSpec.describe PgEventstore::CommandHandlers::SubscriptionRunnersCommands do
     PgEventstore::SubscriptionRunner.new(
       stats: PgEventstore::SubscriptionHandlerPerformance.new,
       events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-      subscription: SubscriptionsHelper.create_with_connection(name: 'Subscr1'),
-      position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-        config_name: :default, filter_options: {}
-      )
+      subscription: SubscriptionsHelper.create_with_connection(name: 'Subscr1')
     )
   end
   let(:runner2) do
     PgEventstore::SubscriptionRunner.new(
       stats: PgEventstore::SubscriptionHandlerPerformance.new,
       events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-      subscription: SubscriptionsHelper.create_with_connection(name: 'Subscr2'),
-      position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-        config_name: :default, filter_options: {}
-      )
+      subscription: SubscriptionsHelper.create_with_connection(name: 'Subscr2')
     )
   end