+++ spec/pg_eventstore/subscriptions/subscription_feeder_spec.rb
@@ -29,20 +29,14 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.init_with_connection,
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.init_with_connection
       )
     end
     let(:subscription_runner2) do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.init_with_connection(name: 'Bar'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.init_with_connection(name: 'Bar')
       )
     end
     let(:setup_subscription_runners) do
@@ -145,20 +139,14 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.init_with_connection,
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.init_with_connection
       )
     end
     let(:subscription_runner2) do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.init_with_connection(name: 'Bar'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.init_with_connection(name: 'Bar')
       )
     end
     let(:setup_subscription_runners) do
@@ -316,20 +304,14 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: subscription1,
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: subscription1
       )
     end
     let(:subscription_runner2) do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: subscription2,
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: subscription2
       )
     end
     let(:subscription_cmd_queries) { PgEventstore::SubscriptionCommandQueries.new(PgEventstore.connection) }
@@ -417,10 +399,7 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
         events_processor: PgEventstore::EventsProcessor.new(
           proc { |event| processed_events.push(event) }, graceful_shutdown_timeout: 5
         ),
-        subscription: SubscriptionsHelper.init_with_connection(set: set_name),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.init_with_connection(set: set_name)
       )
     end
     let(:processed_events) { [] }
@@ -532,10 +511,7 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
         events_processor: PgEventstore::EventsProcessor.new(
           proc { |raw_event| processed_events1.push(raw_event) }, graceful_shutdown_timeout: 5
         ),
-        subscription: SubscriptionsHelper.create_with_connection(options: { filter: { event_types: ['Bar'] } }),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: { event_types: ['Bar'] }
-        )
+        subscription: SubscriptionsHelper.create_with_connection(options: { filter: { event_types: ['Bar'] } })
       )
     end
     let(:subscription_runner2) do
@@ -546,9 +522,6 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
         ),
         subscription: SubscriptionsHelper.create_with_connection(
           name: 'sub2', options: { filter: { event_types: ['Baz'] } }
-        ),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: { event_types: ['Baz'] }
         )
       )
     end
@@ -609,20 +582,14 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.create_with_connection(name: 'Foo'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.create_with_connection(name: 'Foo')
       )
     end
     let(:subscription_runner2) do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.create_with_connection(name: 'Bar'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.create_with_connection(name: 'Bar')
       )
     end
     let(:subscription_cmd_queries) { PgEventstore::SubscriptionCommandQueries.new(PgEventstore.connection) }
@@ -692,20 +659,14 @@ RSpec.describe PgEventstore::SubscriptionFeeder do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.create_with_connection(name: 'Foo'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.create_with_connection(name: 'Foo')
       )
     end
     let(:subscription_runner2) do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(proc {}, graceful_shutdown_timeout: 5),
-        subscription: SubscriptionsHelper.create_with_connection(name: 'Bar'),
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: SubscriptionsHelper.create_with_connection(name: 'Bar')
       )
     end
     let(:set_cmd_queries) { PgEventstore::SubscriptionsSetCommandQueries.new(PgEventstore.connection) }