+++ spec/pg_eventstore/subscriptions/subscription_runner_commands/restore_spec.rb
@@ -31,10 +31,7 @@ RSpec.describe PgEventstore::SubscriptionRunnerCommands::Restore do
       PgEventstore::SubscriptionRunner.new(
         stats: PgEventstore::SubscriptionHandlerPerformance.new,
         events_processor: PgEventstore::EventsProcessor.new(handler, graceful_shutdown_timeout: 5),
-        subscription: subscription,
-        position_evaluation: PgEventstore::SubscriptionPositionEvaluation.new(
-          config_name: :default, filter_options: {}
-        )
+        subscription: subscription
       )
     end
     let(:handler) { proc {} }