+++ spec/pg_eventstore/subscriptions/subscription_runner_spec.rb
@@ -5,8 +5,7 @@ RSpec.describe PgEventstore::SubscriptionRunner do
     described_class.new(
       stats: stats,
       events_processor: events_processor,
-      subscription: subscription,
-      position_evaluation: position_evaluation
+      subscription: subscription
     )
   end
   let(:stats) { PgEventstore::SubscriptionHandlerPerformance.new }
@@ -14,9 +13,6 @@ RSpec.describe PgEventstore::SubscriptionRunner do
     PgEventstore::EventsProcessor.new(handler, graceful_shutdown_timeout: 5)
   end
   let(:subscription) { SubscriptionsHelper.create_with_connection(name: 'Foo') }
-  let(:position_evaluation) do
-    PgEventstore::SubscriptionPositionEvaluation.new(config_name: :default, filter_options: {})
-  end
   let(:handler) { proc {} }
 
   describe '#next_chunk_query_opts' do
@@ -223,8 +219,7 @@ RSpec.describe PgEventstore::SubscriptionRunner do
             filter: { event_types: ['Foo'] },
             resolve_link_tos: true,
             from_position: 124,
-            max_count: described_class::INITIAL_EVENTS_PER_CHUNK,
-            to_position: nil
+            max_count: described_class::INITIAL_EVENTS_PER_CHUNK
           )
         )
       end
@@ -234,16 +229,31 @@ RSpec.describe PgEventstore::SubscriptionRunner do
   describe '#time_to_feed?' do
     subject { instance.time_to_feed? }
 
-    context 'when it is not time to feed' do
+    context 'when #estimate_events_number is greater than zero' do
+      context 'when last feed was more than Subscription#chunk_query_interval seconds ago' do
+        before do
+          subscription.update(last_chunk_fed_at: Time.now.utc - subscription.chunk_query_interval)
+        end
+
+        it { is_expected.to eq(true) }
+      end
+
+      context 'when last feed was less than Subscription#chunk_query_interval seconds ago' do
+        before do
+          subscription.update(last_chunk_fed_at: Time.now.utc)
+        end
+
+        it { is_expected.to eq(false) }
+      end
+    end
+
+    context 'when #estimate_events_number is zero' do
       before do
-        subscription.update(last_chunk_fed_at: Time.now.utc)
+        subscription.update(last_chunk_fed_at: Time.now.utc - subscription.chunk_query_interval)
+        allow(instance).to receive(:estimate_events_number).and_return(0)
       end
 
       it { is_expected.to eq(false) }
-    end
-
-    context 'when it is time to feed' do
-      it { is_expected.to eq(true) }
     end
   end