+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+- **Breaking change**: `pg_eventstore` now requires [pg_cron](https://github.com/citusdata/pg_cron) extension
+- **Breaking change**: `pg_eventstore` now requires PostgreSQL v16+
+- Greatly decreased the number of connections, used by `pg_eventstore` subscriptions
+- `Client#multiple` method now accepts `read_only` keyword argument. When it is set to true - transaction is run in read-only mode 
+
 ## [1.13.4]
 
 - Fix subscriptions potentially skipping events when multiple events are appended in concurrent transactions